requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
Django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
lxml==4.9.2
outcome==1.2.0
psycopg2==2.9.5
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2022.7.1
PyYAML==6.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.14
wsproto==1.2.0
XlsxWriter==3.0.7
